backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
passlib[bcrypt]
bcrypt
python-multipart
PyPDF2
python-docx
groq
email-validator